Repair other tests and configurations that need to initialize both the v2 and v3 support environments.
--HG--
branch : feature/docsrc

+++ sim-support/src/main/groovy/gov/nist/hit/ds/simSupport/utilities/SimUtils.groovy
@@ -173,7 +173,7 @@ class SimUtils {
     }
 
     static def close(SimHandle simHandle) {
-        assert simHandle.open
+        if (!simHandle.open) return
         simHandle.open = false
         simHandle.event.flushAll()
     }
